Separates tracks from strudel imports to prevent unit test issues
