+++ js/audio.js
@@ -1,5 +1,8 @@
-import { startMusic, Track, SpecialTrack } from "./audio-music.js";
-export { Track, SpecialTrack } from "./audio-music.js";
+import { loadTrack } from "./strudel.js";
+import { Track, SpecialTrack } from "./tracks.js";
+
+// Re-exporting for easier imports elsewhere
+export { Track, SpecialTrack } from "./tracks.js";
 
 const audioContext = new (window.AudioContext || /** @type {any} */ (window).webkitAudioContext)();
 
@@ -16,7 +19,7 @@ const currentTrack = {
 
 export function setMusicOptions(options) {
     currentTrack.options = { ...currentTrack.options, ...options };
-    startMusic(currentTrack.track, currentTrack.options);
+    loadTrack(currentTrack.track, currentTrack.options);
 }
 
 export function setMusicTrack(trackNameOrNumber, options) {
